+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `filesToIgnore` option to the configs.
 
+### Changed
+
+- Validate extra keys on region specific files.
+
 ## [2.3.0] - 2019-03-25
 
 ### Added